backend/economic_platform/countries/serializers.py
from rest_framework import serializers
from .models import Country, EconomicIndicator

class CountrySerializer(serializers.ModelSerializer):
    class Meta:
        model = Country
        fields = '__all__'

class EconomicIndicatorSerializer(serializers.ModelSerializer):
    class Meta:
        model = EconomicIndicator
        fields = '__all__' 